+++ apps/happy-blocks/src/support-content/README.md
@@ -1,0 +1,3 @@
+Content embed comes in two flavors:
+- `happy-blocks/support-page` for embedding support pages aka WordPress Guide
+- `happy-blocks/forum` for embedding WordPress forum topics
