+++ apps/happy-blocks/src/support-content/embed-placeholder.tsx
