+++ apps/happy-blocks/src/support-content/inline-skeleton.tsx
