+++ apps/happy-blocks/src/support-content/view.scss
