+++ apps/happy-blocks/src/view.scss
@@ -1,1 +1,1 @@
-@import "./support-page/view";
+@import "./support-content/view";
